Add hook to App.resolve to patch objects before returning them
Path Item with $ref would be merged into one object. The usage of those
merged object is intuitive because you don't have to traverse those '$ref'
path-items to find the correct operation object to call.

However, it's problematic to 'dump' those objects because they are changed
and look different from spec.

In OpenAPI 3.0.0, I still merge them into one object, however, I choose
to keep the merged object to an internal field. This way, the result of dumping
is correct.

However, to use such object, you need to call pyopenapi.utils.final before using it.
So I provide a hook to call this function automatically in App.resolve. This way makes
the usage of pyopenapi and pyswagger won't change when migrating to OpenAPI 3.0.0.